+++ client_mod/mods-unpacked/RampagingHippy-Archipelago/singletons/ap_websocket_connection.gd
@@ -1,14 +1,5 @@
 extends Node
 class_name ApWebSocketConnection
-# Hard-code mod name to avoid cyclical dependency
-var LOG_NAME = "RampagingHippy-Archipelago/AP Client"
-
-# The client handles connecting to the server, and the peer handles sending/receiving
-# data after connecting. We set the peer in the "_on_connection_established" callback,
-# and clear it in the "_on_connection_closed" callback.
-var _client = WebSocketClient.new()
-var _peer: WebSocketPeer
-var _url: String
 
 enum State {
 	STATE_CONNECTING = 0
@@ -16,23 +7,26 @@ enum State {
 	STATE_CLOSING = 2
 	STATE_CLOSED = 3
 }
-signal connection_state_changed
+
+# Hard-code mod name to avoid cyclical dependency
+const LOG_NAME = "RampagingHippy-Archipelago/ap_websocket_connection"
+const _DEFAULT_PORT = 38281
+const _CONNECT_TIMEOUT_SECONDS = 5
+
+# The client handles connecting to the server, and the peer handles sending/receiving
+# data after connecting. We set the peer in the "_on_connection_established" callback,
+# and clear it in the "_on_connection_closed" callback.
+var _client: WebSocketClient
+var _peer: WebSocketPeer
+var _url: String
+var _waiting_to_connect_to_server = null
 
 var connection_state = State.STATE_CLOSED
 
-enum ClientStatus {
-## This is the Client States enum as documented in AP->network protocol
-CLIENT_UNKOWN = 0
-CLIENT_CONNECTED = 5
-CLIENT_READY = 10
-CLIENT_PLAYING = 20
-CLIENT_GOAL = 30
-}
-
-signal item_received
-signal on_room_info
-signal on_connected
-signal on_connection_refused
+signal connection_state_changed
+signal on_connected(connection_data)
+signal on_connection_refused(refused_reason)
+signal on_room_info(room_info)
 signal on_received_items
 signal on_location_info
 signal on_room_update
@@ -43,49 +37,84 @@ signal on_invalid_packet
 signal on_retrieved
 signal on_set_reply
 
+signal _stop_waiting_to_connect(success)
+
 func _ready():
-	# Connect base signals to get notified of connection open, close, and errors.
-	_client.connect("connection_closed", self, "_on_connection_closed")
-	_client.connect("connection_error", self, "_on_connection_error")
-	_client.connect("connection_established", self, "_on_connection_established")
-	_client.connect("data_received", self, "_on_data_received")
-	# Increase max buffer size to accommodate larger payloads. The defaults are:
-	#   - Max in/out buffer = 64 KB
-	#   - Max in/out packets = 1024 
-	# We increase the in buffer to 256 KB because some messages we receive are too large
-	# for 64. The other defaults are fine though.
-	_client.set_buffers(256, 1024, 64, 1024)
-	
 	# Always process so we don't disconnect if the game is paused for too long.
 	pause_mode = Node.PAUSE_MODE_PROCESS
-	
-# Public API
-func connect_to_multiworld(multiworld_url: String):
-	if connection_state == State.STATE_OPEN:
-		return
-	_set_connection_state(State.STATE_CONNECTING)
-	# Try to connect with SSL first. If this doesn't work then the _on_connection_error
-	# callback will try again without SSL.
-	_url = "wss://%s" % multiworld_url
-	ModLoaderLog.info("Connecting to %s" % _url, LOG_NAME)
-	var err = _client.connect_to_url(_url)
-	if not err:
-		# Start processing to poll the connection for data
-		set_process(true)
 
-func connected_to_multiworld() -> bool:
+# Public API
+func connect_to_server(server: String) -> bool:
+	if connection_state == State.STATE_OPEN:
+		return true
+	_set_connection_state(State.STATE_CONNECTING)
+
+	# TODO: It would be nice to move a lot of this into a separate factory class so we
+	# didn't have to duplicate code for the WSS and WS cases, and so we don't have
+	# methods that all only called a specific times which could otherwise be discarded.
+	# But, my one attempt at doing so led to weird cases where the WS client didn't
+	# correctly perform the handshake with the server.
+
+	# Use the default Archipelago port if not included in the URL
+	var port_check_pattern = RegEx.new()
+	port_check_pattern.compile(":(\\d+)$")
+	var server_has_port = port_check_pattern.search(server)
+	if not server_has_port:
+		server = "%s:%d" % [server, _DEFAULT_PORT]
+
+	# Try to connect with SSL first
+	var wss_url = "wss://%s" % [server]
+
+	# Create a timeout to trigger the done waiting signal if we take too long
+	_init_client()
+	_waiting_to_connect_to_server = wss_url
+	_make_connection_timeout(wss_url)
+	var _result = _client.connect_to_url(wss_url) # Return value is useless
+
+	var wss_success = yield (self, "_stop_waiting_to_connect")
+	_waiting_to_connect_to_server = null
+
+	var ws_success = false
+	if not wss_success:
+		# We don't have any info on why the connection failed (thanks Godot), so we
+		# assume it was because the server doesn't support SSL. So, try connecting using
+		# "ws://" instead.
+		ModLoaderLog.info("Connecting with WSS failed, trying WS.", LOG_NAME)
+		var ws_url = "ws://%s" % [server]
+		_init_client()
+		_waiting_to_connect_to_server = ws_url
+		_make_connection_timeout(ws_url)
+		_result = _client.connect_to_url(ws_url)
+
+		ws_success = yield (self, "_stop_waiting_to_connect")
+		_waiting_to_connect_to_server = null
+		if ws_success:
+			_url = ws_url
+	else:
+		_url = wss_url
+
+	if wss_success or ws_success:
+		_peer = _client.get_peer(1)
+		_peer.set_write_mode(WebSocketPeer.WRITE_MODE_TEXT)
+		_set_connection_state(State.STATE_OPEN)
+		ModLoaderLog.info("Connected to multiworld %s." % _url, LOG_NAME)
+	
+	return wss_success or ws_success
+
+func connected_to_server() -> bool:
 	return connection_state == State.STATE_OPEN
 	
-func disconnect_from_multiworld():
+func disconnect_from_server():
 	if connection_state == State.STATE_CLOSED:
 		return
 	_set_connection_state(State.STATE_CLOSING)
+	# The "connection_closed" signal handler will take care of cleanup
 	_client.disconnect_from_host()
 
-func send_connect(game: String, user: String, password: String = "", slot_data: bool = true):
+func send_connect(game: String, user: String, password: String="", slot_data: bool=true):
 	_send_command({
-		"cmd": "Connect", 
-		"game": game, 
+		"cmd": "Connect",
+		"game": game,
 		"name": user,
 		"password": password,
 		"uuid": "Godot %s: %s" % [game, user], # TODO: What do we need here? We can't generate an actual UUID in 3.5
@@ -166,38 +195,27 @@ func set_notify(keys: Array):
 		"keys": keys,
 	})
 
-# WebsocketClient callbacks
-func _send_command(args: Dictionary):
-	ModLoaderLog.info("Sending %s command" % args["cmd"], LOG_NAME)
-	var command_str = JSON.print([args])
-	var _result = _peer.put_packet(command_str.to_ascii())
-
-func _on_connection_closed(was_clean = false):
-	_set_connection_state(State.STATE_CLOSED)
-	ModLoaderLog.info("AP connection closed, clean: %s" % was_clean, LOG_NAME)
-	_peer = null
-	set_process(false)
-
-func _on_connection_established(_proto = ""):
-	_set_connection_state(State.STATE_OPEN)
-	_peer = _client.get_peer(1)
-	_peer.set_write_mode(WebSocketPeer.WRITE_MODE_TEXT)
-	ModLoaderLog.info("Connected to multiworld %s." % _url, LOG_NAME)
+# WebSocketClient callbacks
+func _on_connection_established(_proto=""):
+	# We succeeded, stop waiting and tell the caller.
+	ModLoaderLog.debug("Successfully connected.", LOG_NAME)
+	emit_signal("_stop_waiting_to_connect", true)
 
 func _on_connection_error():
-	if _url.begins_with("wss://"):
-		# We don't have any info on why the connection failed, so we assume it wsa
-		# because the server doesn't support SSL. So, try connecting using "ws://"
-		# instead.
-		ModLoaderLog.debug("Connecting to multiworld %s failed, trying again using 'ws://'." % _url, LOG_NAME)
-		_url = _url.replace("wss://", "ws://")
-		_client.connect_to_url(_url)
-	else:
-		# Tried both options, error out now
-		_set_connection_state(State.STATE_CLOSED)
-		ModLoaderLog.info("Failed to connect to multiworld %s." % _url, LOG_NAME)
+	# We failed, stop waiting and tell the caller.
+	ModLoaderLog.debug("Error connecting.", LOG_NAME)
+	emit_signal("_stop_waiting_to_connect", false)
+
+func _on_connection_closed(was_clean=false):
+	_set_connection_state(State.STATE_CLOSED)
+	ModLoaderLog.info("AP connection closed, clean: %s." % was_clean, LOG_NAME)
+	_peer = null
 
 func _on_data_received():
+	if self._peer == null:
+		# Rare case where we have an dead connection to a server that suddenly
+		# becomes active. Just ignore because we're not setup for it.
+		return
 	var received_data_str = _peer.get_packet().get_string_from_utf8()
 	var received_data = JSON.parse(received_data_str)
 	if received_data.result == null:
@@ -207,8 +225,46 @@ func _on_data_received():
 		_handle_command(command)
 
 # Internal plumbing
+func _send_command(args: Dictionary):
+	ModLoaderLog.info("Sending %s command" % args["cmd"], LOG_NAME)
+	var command_str = JSON.print([args])
+	var _result = _peer.put_packet(command_str.to_ascii())
+
+func _init_client():
+	if self._client != null:
+		# Disconnect signals from the old client reference
+		self._client.disconnect("connection_closed", self, "_on_connection_closed")
+		self._client.disconnect("data_received", self, "_on_data_received")
+		self._client.disconnect("connection_established", self, "_on_connection_established")
+		self._client.disconnect("connection_error", self, "_on_connection_error")
+	self._client = WebSocketClient.new()
+	
+	# Connect base signals to get notified of connection open, close, and errors.
+	var _result = self._client.connect("connection_closed", self, "_on_connection_closed")
+	_result = self._client.connect("data_received", self, "_on_data_received")
+	_result = self._client.connect("connection_established", self, "_on_connection_established")
+	_result = self._client.connect("connection_error", self, "_on_connection_error")
+	
+	# Increase max buffer size to accommodate AP's larger payloads. The defaults are:
+	#   - Max in/out buffer = 64 KB
+	#   - Max in/out packets = 1024 
+	# We increase the in buffer to 256 KB because some messages we receive are too large
+	# for 64. The other defaults are fine though.
+	_result = _client.set_buffers(256, 1024, 64, 1024)
+
+	self._peer = null
+
+func _make_connection_timeout(for_url: String):
+	yield (get_tree().create_timer(_CONNECT_TIMEOUT_SECONDS), "timeout")
+	if _waiting_to_connect_to_server == for_url:
+		# We took to long, stop waiting and tell the called we failed.
+		_waiting_to_connect_to_server = false
+		ModLoaderLog.debug("Timed out trying to connect.", LOG_NAME)
+		emit_signal("_stop_waiting_to_connect", false)
+
 func _set_connection_state(state):
-	ModLoaderLog.info("AP connection state changed to: %d" % state, LOG_NAME)
+	var state_name = State.keys()[state]
+	ModLoaderLog.info("AP connection state changed to: %s." % state_name, LOG_NAME)
 	connection_state = state
 	emit_signal("connection_state_changed", connection_state)
 
@@ -255,4 +311,5 @@ func _handle_command(command: Dictionary):
 
 func _process(_delta):
 	# Only run when the connection the the server is not closed.
-	_client.poll()
+	if _client != null:
+		_client.poll()