+++ client_mod/mods-unpacked/RampagingHippy-Archipelago/ui/menus/pages/menu_ap_connect.gd
@@ -2,6 +2,10 @@ extends MarginContainer
 
 signal back_button_pressed
 
+var _ap_icon_connected = preload("res://mods-unpacked/RampagingHippy-Archipelago/ap_button_icon_connected.png")
+var _ap_icon_disconnected = preload("res://mods-unpacked/RampagingHippy-Archipelago/ap_button_icon_disconnected.png")
+var _ap_icon_error =  preload("res://mods-unpacked/RampagingHippy-Archipelago/ap_button_icon_error.png")
+
 onready var _connect_button: Button = $"VBoxContainer/ConnectButton"
 onready var _disconnect_button: Button = $"VBoxContainer/DisconnectButton"
 onready var _connect_status_label: Label = $"VBoxContainer/ConnectStatusLabel"
@@ -9,9 +13,13 @@ onready var _connect_error_label: Label = $"VBoxContainer/ConnectionErrorLabel"
 onready var _host_edit: LineEdit = $"VBoxContainer/CenterContainer/GridContainer/HostEdit"
 onready var _player_edit: LineEdit = $"VBoxContainer/CenterContainer/GridContainer/PlayerEdit"
 onready var _password_edit: LineEdit = $"VBoxContainer/CenterContainer/GridContainer/PasswordEdit"
+onready var _status_texture: TextureRect = $"VBoxContainer/StatusTexture"
 
-onready var _ap_websocket_connection
-onready var _ap_client
+onready var _ap_session
+
+const _MAX_ANGLE_DEGREES = 360
+const _STATUS_TEXTURE_ROTATION_SPEED_DEGREES_PER_SECOND = 360
+var _animate_status_texture: bool = false
 
 func init():
 	# Needed to make the scene switch in title_screen_menus happy.
@@ -19,65 +27,130 @@ func init():
 
 func _ready():
 	var mod_node = get_node("/root/ModLoader/RampagingHippy-Archipelago")
-	_ap_websocket_connection = mod_node.ap_websocket_connection
-	_ap_client = mod_node.brotato_ap_client
-	_ap_websocket_connection.connect("connection_state_changed", self, "_on_connection_state_changed")
-	_ap_client.connect("on_connection_refused", self, "_on_connection_refused")
-	_on_connection_state_changed(_ap_websocket_connection.connection_state)
+	
+	_ap_session = mod_node.ap_player_session
+	_ap_session.connect("connection_state_changed", self, "_on_connection_state_changed")
+	
 
 #func _input(_event):
 #	if get_tree().current_scene.name == self.name && Input.is_key_pressed(KEY_ENTER):
 #		_on_ConnectButton_pressed()
 
-func _on_connection_state_changed(new_state, connection_refused: bool = false):
+func _on_connection_state_changed(new_state: int, error: int=0):
+	# See ConnectState enum in ap_player_session.gd
 	match new_state:
 		0:
+			# Disconnected
+			_connect_status_label.text = "Disconnected"
+		1:
 			# Connecting
 			_connect_status_label.text = "Connecting"
-		1:
-			# Open
-			_connect_status_label.text = "Connected"
 		2:
-			# Closing
+			# Disconnecting
 			_connect_status_label.text = "Disconnecting"
 		3:
-			# Closed
-#			_connect_button.text = "Connect"
-			_connect_status_label.text = "Disconnected"
-	var show_connect_button = new_state == 3 or (new_state == 1 and connection_refused)
-	var show_disconnect_button = not show_connect_button
-	_connect_button.visible = show_connect_button
-	_disconnect_button.visible = show_disconnect_button
-	# _connect_error_label.visible = show_disconnect_button
+			# Connected to server
+			_connect_status_label.text = "Connected to server"
+		4:
+			# Connected to multiworld
+			_connect_status_label.text = "Connected to multiworld"
 
-func _on_connection_refused(reasons: Array):
-	#TODO: handle multiple errors
-	var reason_string
-	match reasons[0]:
-		"InvalidSlot":
-			reason_string = "Invalid slot: %s." % _ap_client.player
-		"InvalidGame":
-			reason_string = "Slot for %s is not a Brotato game." % _ap_client.player
-		"IncompatibleVersion":
-			reason_string = "Version mismatch."
-		"InvalidPassword":
-			reason_string = "Invalid password."
-		"InvalidItemsHandling":
-			reason_string = "Invalid items handling (oops)."
-	_connect_error_label.text = "Connection Refused: " + reason_string
-	# _on_connection_state_changed(1, true)
+	# Allow connecting if disconnected or connected to the server but not the multiworld
+	_connect_button.disabled = (
+		new_state == ApPlayerSession.ConnectState.CONNECTED_TO_MULTIWORLD or
+		new_state == ApPlayerSession.ConnectState.DISCONNECTING
+	)
+	if _connect_button.disabled and _connect_button.has_focus():
+		# Disabled buttons having focus look ugly and don't make sense.
+		_connect_button.release_focus()
+
+	# Allow disconnecting if connected to the server and/or multiworld
+	_disconnect_button.disabled = (
+		new_state == ApPlayerSession.ConnectState.DISCONNECTED or
+		new_state == ApPlayerSession.ConnectState.DISCONNECTED or
+		 # TODO: Remove below if we figure out how to cancel the connection process.
+		new_state == ApPlayerSession.ConnectState.CONNECTING
+	)
+
+	if _disconnect_button.disabled and _disconnect_button.has_focus():
+		_disconnect_button.release_focus()
+
+	if new_state == ApPlayerSession.ConnectState.CONNECTED_TO_MULTIWORLD:
+		_status_texture.texture = _ap_icon_connected
+	elif error != 0:
+		_status_texture.texture = _ap_icon_error
+	else:
+		_status_texture.texture = _ap_icon_disconnected
+		
+	if new_state == ApPlayerSession.ConnectState.CONNECTING:
+		_animate_status_texture = true
+	else:
+		_animate_status_texture = false
+		_status_texture.rect_rotation = 0
 	
+	if error != 0:
+		_set_error(error)
+	else:
+		_clear_error()
+
+func _set_error(error_reason: int):
+	# See ConnectResult enum in ap_player_session.gd
+	var error_text: String
+	match error_reason:
+		1:
+			error_text = "Failed to connect to the server"
+		2:
+			error_text = "Need to set player name before connecting"
+		3:
+			error_text = "Client needs to set game name before connecting"
+		4:
+			error_text = "Invalid server name"
+		5:
+			error_text = "AP: Invalid player name"
+		6:
+			error_text = "AP: Invalid game"
+		7:
+			error_text = "AP: Incompatible versions"
+		8:
+			error_text = "AP: Invalid or missing password"
+		9:
+			error_text = "AP: Invalid items handling"
+		9:
+			error_text = "AP: Failed to connect (unknown error)"
+		_:
+			error_text = "Unknown error"
+
+	_connect_error_label.visible = true
+	_connect_error_label.text = error_text
+
+func _clear_error():
+	_connect_error_label.visible = false
+	_connect_error_label.text = ""
 
 func _on_ConnectButton_pressed():
-	var url = _host_edit.text
-	_ap_client.player = _player_edit.text
-	_ap_client.password = _password_edit.text
-	_ap_websocket_connection.connect_to_multiworld(url)
-
+	_ap_session.server = _host_edit.text
+	_ap_session.player = _player_edit.text
+ 
+	# Fire and forget this coroutine call, signal handlers will take care of the rest.
+	_ap_session.connect_to_multiworld(_password_edit.text)
 
 func _on_BackButton_pressed():
 	emit_signal("back_button_pressed")
 
-
 func _on_DisconnectButton_pressed():
-	_ap_websocket_connection.disconnect_from_multiworld()
+	_ap_session.disconnect_from_multiworld()
+
+func _reset_status_texture():
+	_status_texture.set_rotation(0)
+
+func _process(delta):
+	if _animate_status_texture:
+		# Set status texture to pivot around its center instead of its top-left corner.
+		# Do this every frame in case the screen size changes.
+		_status_texture.rect_pivot_offset = _status_texture.rect_size / 2
+		var new_angle = _status_texture.rect_rotation + (delta * _STATUS_TEXTURE_ROTATION_SPEED_DEGREES_PER_SECOND)
+		if new_angle > _MAX_ANGLE_DEGREES:
+			# Rotation goes from -360 to 360 degrees
+			new_angle -= _MAX_ANGLE_DEGREES * 2
+		_status_texture.rect_rotation = new_angle
+